Сделал так, чтобы можно было удалять сообщение с дизлайками (до этого была ошибка об ограничениях внешнего ключа при простом удалении). Поправил сервисы

+++ src/main/java/org/forum/main/services/implementations/StatisticsServiceImpl.java
@@ -86,18 +86,18 @@ public class StatisticsServiceImpl implements StatisticsService {
     }
 
     @Override
-    public long messagesCount() {
-        return messageRepository.count();
+    public int messagesCount() {
+        return (int) messageRepository.count();
     }
 
     @Override
-    public long likesCount() {
-        return likeRepository.count();
+    public int likesCount() {
+        return (int) likeRepository.count();
     }
 
     @Override
-    public long dislikesCount() {
-        return dislikeRepository.count();
+    public int dislikesCount() {
+        return (int) dislikeRepository.count();
     }
 
     @Override
