+++ src/main/java/org/forum/main/services/interfaces/StatisticsService.java
@@ -23,11 +23,11 @@ public interface StatisticsService {
 
     int topicsCount();
 
-    long messagesCount();
+    int messagesCount();
 
-    long likesCount();
+    int likesCount();
 
-    long dislikesCount();
+    int dislikesCount();
 
     int bansCount();
